Converted all int to long to support large article id

+++ server/src/solver/solver/bfs.h
@@ -5,16 +5,16 @@
 
 // File d'attente (queue) pour le parcours en largeur
 typedef struct {
-    int *items;     // tableau des éléments
-    int front;      // index du premier élément
-    int rear;       // index du dernier élément
-    int capacity;   // capacité maximale de la file
+    long *items;     // tableau des éléments
+    long front;      // index du premier élément
+    long rear;       // index du dernier élément
+    long capacity;   // capacité maximale de la file
 } Queue;
 
 // Création et gestion de la file
-Queue* createQueue(int capacity);
+Queue* createQueue(long capacity);
 int   isEmpty(Queue* q);
-void  enqueue(Queue* q, int v);
+void  enqueue(Queue* q, long v);
 int   dequeue(Queue* q);
 void  freeQueue(Queue* q);
 
@@ -22,12 +22,12 @@ void  freeQueue(Queue* q);
 // parent : tableau des prédécesseurs pour chaque sommet
 // depart : indice du sommet de départ
 // objectif : indice du sommet d'arrivée
-void print_path(Graphe* graphe, int *parent, int depart, int objectif);
+void print_path(Graphe* graphe, long *parent, long depart, long objectif);
 
 // Fonction principale de l'algorithme BFS
 // graphe   : pointeur vers le graphe
 // depart   : indice du sommet de départ
 // objectif : indice du sommet d'arrivée (arrêt anticipé possible)
-void bfs(Graphe* graphe, int depart, int objectif);
+void bfs(Graphe* graphe, long depart, long objectif);
 
 #endif // BFS_H
